Section 21: Pagination

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const { deleteFile } = require('../util/file')
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -15,7 +17,20 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
 
-  const { title, description, price, imageUrl } = req.body
+  const { title, description, price } = req.body;
+  const image = req.file;
+  if (!image) {
+    return res
+      .status(422)
+      .render('admin/edit-product', {
+        path: '/admin/edit-product',
+        pageTitle: 'Edit Product',
+        editing: false,
+        hasError: true,
+        errorMessage: 'attatched file is not an image',
+        product: { title, description, price }
+      });
+  }
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res
@@ -26,9 +41,11 @@ exports.postAddProduct = (req, res, next) => {
         editing: false,
         hasError: true,
         errorMessage: errors.array()[0].msg,
-        product: { title, description, price, imageUrl }
+        product: { title, description, price }
       });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
@@ -45,7 +62,10 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      console.log(err);
+      console.log('here')
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error)
     });
 };
 
@@ -69,15 +89,20 @@ exports.getEditProduct = (req, res, next) => {
         product: product
       });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error)
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
+
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -89,7 +114,7 @@ exports.postEditProduct = (req, res, next) => {
         editing: true,
         hasError: true,
         errorMessage: errors.array()[0].msg,
-        product: { title: updatedTitle, description: updatedDesc, price: updatedPrice, imageUrl: updatedImageUrl, _id: prodId }
+        product: { title: updatedTitle, description: updatedDesc, price: updatedPrice, _id: prodId }
       });
   }
 
@@ -101,13 +126,20 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error)
+    });
 };
 
 exports.getProducts = (req, res, next) => {
@@ -122,15 +154,30 @@ exports.getProducts = (req, res, next) => {
         path: '/admin/products'
       });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error)
+    });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'))
+      }
+      deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error)
+    });
 };
